+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.2.0 (TBA)
 
 * Added required `messaging_type` property (https://developers.facebook.com/docs/messenger-platform/send-messages#messaging_types) to all send operations.
+* Added support for the open graph template
 * Upgraded RestFB dependency to latest 1.48.0
 * Upgraded Spring Boot Starter to latest Spring Boot 1.5.8
 * Updated default Facebook API version to 2.11